@@ -4,9 +4,9 @@
   "private": true,
   "dependencies": {
     "@atlaskit/atlassian-navigation": "^2.2.6",
-    "@atlaskit/avatar": "^21.0.0",
+    "@atlaskit/avatar": "^21.1.1",
     "@atlaskit/banner": "^11.6.1",
-    "@atlaskit/breadcrumbs": "^11.6.2",
+    "@atlaskit/breadcrumbs": "^11.7.3",
     "@atlaskit/button": "^16.3.5",
     "@atlaskit/checkbox": "^12.3.18",
     "@atlaskit/comment": "^10.5.2",
@@ -23,7 +23,7 @@
     "@atlaskit/progress-bar": "^0.5.8",
     "@atlaskit/section-message": "^6.1.14",
     "@atlaskit/spinner": "^15.1.14",
-    "@atlaskit/select": "^15.6.0",
+    "@atlaskit/select": "^15.7.5",
     "@atlaskit/textarea": "^4.3.8",
     "@atlaskit/textfield": "^5.1.12",
     "@atlaskit/tree": "^8.6.1",
@@ -42,11 +42,6 @@
     "@typecell-org/parsers": "^0.0.3",
     "@typecell-org/common": "^0.0.3",
     "@typecell-org/engine": "^0.0.3",
-    "@types/markdown-it": "^10.0.3",
-    "@types/node": "^16.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/uuid": "^8.3.0",
     "@typescript/vfs": "^1.3.4",
     "@blocknote/core": "^0.1.0",
     "classnames": "^2.3.1",
@@ -66,14 +61,14 @@
     "prettier": "2.4.1",
     "probe.gl": "^3.1.4",
     "qs": "^6.10.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-avatar": "^3.10.0",
     "react-dnd": "^14.0.2",
     "react-dnd-html5-backend": "^14.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-error-overlay": "^6.0.9",
     "react-github-btn": "^1.2.1",
-    "react-icons": "^4.2.0",
+    "react-icons": "^4.6.0",
     "react-inspector": "^5.1.0",
     "react-intl-next": "npm:react-intl@^5.18.1",
     "remixicon-react": "^1.0.0",
@@ -154,6 +149,10 @@
     "@types/uuid": "^8.3.0",
     "@types/zxcvbn": "^4.4.1",
     "@types/prettier": "^2.6.4",
+    "@types/markdown-it": "^10.0.3",
+    "@types/node": "^16.0.0",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
     "cross-env": "^7.0.3",
     "eslint": "^8.21.0",
     "eslint-config-react-app": "^7.0.1",
