@@ -6,7 +6,7 @@
     "es-module-shims": "1.4.3",
     "lodash": "^4.17.21",
     "mobx": "^6.2.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "vscode-lib": "^0.1.2"
   },
   "devDependencies": {
@@ -15,7 +15,7 @@
     "@types/chai": "^4.3.0",
     "@types/mocha": "^9.1.0",
     "@types/lodash": "^4.14.182",
-    "@types/react": "^17.0.0",
+    "@types/react": "^18.0.25",
     "rimraf": "^3.0.2",
     "typescript": "4.3.2",
     "@playwright/test": "^1.18.1",
